middleware for guarding APIs - init
